feat: move components and add biomejs
