+++ apps/the_monkeys/package.json
@@ -9,10 +9,10 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "format": "prettier --write"
+    "format": "prettier --write ."
   },
   "dependencies": {
-		"@the-monkeys/ui": "workspace:*",
+    "@the-monkeys/ui": "workspace:*",
     "@editorjs/code": "2.9.3",
     "@editorjs/delimiter": "1.4.2",
     "@editorjs/editorjs": "2.29.1",
@@ -35,9 +35,7 @@
     "@radix-ui/react-select": "^2.1.2",
     "@radix-ui/react-separator": "^1.0.3",
     "@radix-ui/react-slot": "^1.1.0",
-    "@radix-ui/react-switch": "^1.0.3",
     "@radix-ui/react-tabs": "^1.0.4",
-    "@radix-ui/react-toast": "^1.1.5",
     "@remixicon/react": "^4.1.1",
     "@tanstack/react-query": "^5.66.3",
     "axios": "^1.6.8",
@@ -72,7 +70,7 @@
   },
   "devDependencies": {
     "@tanstack/react-query-devtools": "^5.66.5",
-		"@the-monkeys/config": "workspace:*",
+    "@the-monkeys/config": "workspace:*",
     "@testing-library/react": "^16.1.0",
     "@testing-library/user-event": "^14.5.2",
     "@trivago/prettier-plugin-sort-imports": "^4.3.0",
