+++ apps/the_monkeys/src/app/auth/components/forms/LoginForm.tsx
@@ -15,12 +15,12 @@ import {
   FormMessage,
 } from '@/components/ui/form';
 import { Input } from '@/components/ui/input';
-import { toast } from '@/components/ui/use-toast';
 import { loginSchema } from '@/lib/schema/auth';
 import { cn } from '@/lib/utils';
 import { login } from '@/services/auth/auth';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { toast } from '@the-monkeys/ui/hooks/use-toast';
 import { useForm } from 'react-hook-form';
 import { z } from 'zod';
