+++ apps/the_monkeys/src/app/auth/google/callback/page.tsx
@@ -10,9 +10,9 @@ import Icon from '@/components/icon';
 import LinksRedirectArrow from '@/components/links/LinksRedirectArrow';
 import { Loader } from '@/components/loader';
 import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert';
-import { toast } from '@/components/ui/use-toast';
 import { googleSSOCallback } from '@/services/auth/auth';
 import { useMutation } from '@tanstack/react-query';
+import { toast } from '@the-monkeys/ui/hooks/use-toast';
 
 import {
   FormHeader,
