+++ apps/the_monkeys/src/app/notifications/components/MarkReadButton.tsx
@@ -2,8 +2,8 @@ import { useState } from 'react';
 
 import Icon from '@/components/icon';
 import { Button } from '@/components/ui/button';
-import { toast } from '@/components/ui/use-toast';
 import axiosInstance from '@/services/api/axiosInstance';
+import { toast } from '@the-monkeys/ui/hooks/use-toast';
 import { mutate } from 'swr';
 import { z } from 'zod';
