+++ apps/the_monkeys/src/app/topics/[topic]/components/TopicFollowButton.tsx
@@ -4,11 +4,11 @@ import { usePathname, useRouter, useSearchParams } from 'next/navigation';
 
 import Icon from '@/components/icon';
 import { Button } from '@/components/ui/button';
-import { toast } from '@/components/ui/use-toast';
 import useAuth from '@/hooks/auth/useAuth';
 import useUser from '@/hooks/user/useUser';
 import axiosInstance from '@/services/api/axiosInstance';
 import { followTopicApi, unfollowTopicApi } from '@/services/user/user';
+import { toast } from '@the-monkeys/ui/hooks/use-toast';
 import { mutate } from 'swr';
 
 interface TopicFollowButtonProps {
