+++ apps/the_monkeys/src/app/topics/explore/components/TopicButton.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 
 import Icon from '@/components/icon';
 import { Button } from '@/components/ui/button';
-import { toast } from '@/components/ui/use-toast';
 import axiosInstance from '@/services/api/axiosInstance';
 import { GetPublicUserProfileApiResponse } from '@/services/profile/userApiTypes';
+import { toast } from '@the-monkeys/ui/hooks/use-toast';
 import { mutate } from 'swr';
 
 interface TopicButtonProps {
