+++ apps/the_monkeys/src/components/blog/buttons/LikeButton.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react';
 
 import Icon from '@/components/icon';
-import { toast } from '@/components/ui/use-toast';
 import { useIsPostLiked } from '@/hooks/user/useLikeStatus';
 import axiosInstance from '@/services/api/axiosInstance';
+import { toast } from '@the-monkeys/ui/hooks/use-toast';
 import { mutate } from 'swr';
 
 export const LikeButton = ({
