+++ apps/the_monkeys/src/components/blog/buttons/RemoveBookmarkButton.tsx
@@ -2,8 +2,8 @@ import { useState } from 'react';
 
 import Icon from '@/components/icon';
 import { Loader } from '@/components/loader';
-import { toast } from '@/components/ui/use-toast';
 import axiosInstance from '@/services/api/axiosInstance';
+import { toast } from '@the-monkeys/ui/hooks/use-toast';
 import { mutate } from 'swr';
 
 export const RemoveBookmarkButton = ({
