+++ apps/the_monkeys/src/components/user/dialogs/ShareProfileDialog.tsx
@@ -14,8 +14,8 @@ import {
   DialogTitle,
   DialogTrigger,
 } from '@/components/ui/dialog';
-import { toast } from '@/components/ui/use-toast';
 import { LIVE_URL } from '@/constants/api';
+import { toast } from '@the-monkeys/ui/hooks/use-toast';
 
 interface ShareProfileDialogProps {
   username: string;
