+++ apps/the_monkeys/src/components/user/dialogs/UpdateUsernameDialog.tsx
@@ -19,13 +19,13 @@ import {
   FormMessage,
 } from '@/components/ui/form';
 import { Input } from '@/components/ui/input';
-import { toast } from '@/components/ui/use-toast';
 import { updateUsername } from '@/lib/schema/settings';
 import { User } from '@/services/models/user';
 import { GetPublicUserProfileApiResponse } from '@/services/profile/userApiTypes';
 import { updateUserName } from '@/services/user/user';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { toast } from '@the-monkeys/ui/hooks/use-toast';
 import { useForm } from 'react-hook-form';
 import { z } from 'zod';
