+++ packages/ui/tsconfig.json
@@ -2,19 +2,9 @@
 	"extends": "@the-monkeys/config/tsconfig.json",
 	"compilerOptions": {
 		"paths": {
-			"@/*": [
-				"./src/*"
-			]
-		},
-		"outDir": "dist"
+			"@/*": ["./src/*"]
+		}
 	},
-	"include": [
-		"**/*.ts",
-		"**/*.js",
-		"**/*.jsx",
-		"**/*.tsx",
-	],
-	"exclude": [
-		"node_modules"
-	]
+	"include": ["**/*.ts", "**/*.js", "**/*.jsx", "**/*.tsx"],
+	"exclude": ["node_modules"]
 }
